% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl2newsmd.R
\name{cl2newsmd}
\alias{cl2newsmd}
\title{Convert ChangeLog/NEWS into NEWS.md}
\usage{
cl2newsmd(log, news = NULL, codify = TRUE, overwrite = TRUE)
}
\arguments{
\item{log}{Character string, path to the ChangeLog or NEWS file to be converted.}

\item{news}{Character string, path to the NEWS.md file to be written.
If \code{NULL}, results are written to \code{stdout()}.}

\item{codify}{Logical,
      whether to try to detect code snippets like function names and mark them up accordingly.}

\item{overwrite}{Logical, whether to overwrite an existing NEWS.md file.}
}
\value{
Either writes a file without a return value, or returns the markdown string vector.
}
\description{
Attempts to translate ASCII ChangeLog (or NEWS) files into NEWS.md files.
}
\details{
This should work for ChangeLog and NEWS files that
\enumerate{
  \item have entries named "Changes in version <version number>" (and optionally a parenthetical YYYY-MM-DD date string afterwards)
  \item have single changes properly itemized, by indentation and then either \code{"o"},
      \code{"-"} or \code{"*"} followed
    by space
  \item optionally have categories as subsections, like "Changed", "Fixed", "Added",
      or "Removed"
}
Any text string that isn't indented and doesn't start with "Changes in version" will likely be treated as a subsection.
The ChangeLog related functions and methods of this package,
      e.g. \code{\link[roxyPackage:initChangeLog]{initChangeLog}}, are
a convenient way to maintain R ChangeLogs in a proper format.

This function uses the internal function \code{tools:::.news_reader_default} on the ChangeLog file provided.
}
\examples{
\dontrun{
cl2newsmd(log="~/myFiles/myRPackage/ChangeLog", news="~/myFiles/myRPackage/inst/NEWS.md")

# dump the results into a character vector
NEWS.object <- cl2newsmd(log="~/myFiles/myRPackage/ChangeLog")
}
}
\seealso{
\code{\link[roxyPackage:initChangeLog]{initChangeLog}},
   \code{\link[roxyPackage:readChangeLog]{readChangeLog}},
   \code{\link[roxyPackage:updateChangeLog]{updateChangeLog}},
   \code{\link[roxyPackage:writeChangeLog]{writeChangeLog}}
}
