% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_icon_fetcher.R
\name{orcid_icon_fetcher}
\alias{orcid_icon_fetcher}
\title{Download ORCID icon}
\usage{
orcid_icon_fetcher(
  repo,
  filename = "orcid.svg",
  url = "https://ndownloader.figshare.com/files/8439047",
  method = "auto",
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{repo}{The repository root directory to save the file.}

\item{filename}{Target filename.}

\item{url}{The URL to fetch the SVG file from.}

\item{method}{The download method to use,
      see \code{\link[utils:download.file]{download.file}}.}

\item{overwrite}{If \code{FALSE} (default) and the target file already exists, it will
not be replaced with a newly downloaded one.}

\item{...}{Further arguments to pass to \code{\link[utils:download.file]{download.file}},
exept for \code{mode} (see Note).}
}
\description{
A simple wrapper for \code{\link[utils:download.file]{download.file}} to download
the official ORCID icon for use in a repository with authenticated ORCID IDs.
}
\note{
The \code{mode} argument is fixed to "wb" (binary) to ensure successful downloads
also on Windows systems.
}
\examples{
\dontrun{
orcid_icon_fetcher("/tmp")
}
}
