% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_file_skeleton.R
\name{package_file_skeleton}
\alias{package_file_skeleton}
\title{Generate package file skeletons}
\usage{
package_file_skeleton(
  name,
  type = c("function", "S4Class", "S4Method", "data")
)
}
\arguments{
\item{name}{Character string, name of the object the file should cover (i.e.,
      a function or class name).}

\item{type}{One of \code{"function"}, \code{"S4Class"}, \code{"S4Method"},
      or \code{"data"}, categorizing the object to be documented.}
}
\description{
Writes a new file with some basic comments and documentation hints,
      according to the type of R object to cover.
}
