% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxyPackage-package.R
\docType{package}
\name{roxyPackage-package}
\alias{roxyPackage}
\alias{roxyPackage-package}
\title{\packageTitle{roxyPackage}}
\description{
\packageDescription{roxyPackage}
}
\details{
The DESCRIPTION file:
\tabular{ll}{
Package: \tab roxyPackage\cr
Type: \tab Package\cr
Version: \tab 0.10-3\cr
Date: \tab 2024-07-23\cr
Depends: \tab R (>= 3.0.0)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab https://reaktanz.de/?c=hacking&s=roxyPackage\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://reaktanz.de/?c=hacking&s=roxyPackage}
  \item Report bugs at \url{https://github.com/unDocUMeantIt/roxyPackage/issues}
}

}
\author{
\packageAuthor{roxyPackage}

Maintainer: \packageMaintainer{roxyPackage}
}
